refactor: merge resource reconcile with ComputeNode resources
